fix ORB arming: == 575 → >= 575 for 60s tick data
The ORB strategy checks est_min == 575 (9:35 AM EST) to arm.
With ticks arriving ~60s apart at :00 seconds, est_min is always
a multiple of 60 (570, 630, 690...) and never exactly 575.
Changed to est_min >= 575 so the first tick at/after 9:35 AM
correctly arms the ORB.

Also cleaned up debug eprintln statements from backtest engine.

+++ libs/nexus/src/ingestion/binance_file.rs
@@ -126,7 +126,7 @@ impl BinanceFileIngestor {
     pub fn new(symbol: &str) -> Self {
         Self {
             symbol: symbol.to_uppercase(),
-            precision: 9,
+            precision: 16,
             anchor_interval: 100,
         }
     }
